renames.. and add more tests
