+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/optimizer/DeterministicLiteralUDFFolding.scala
@@ -26,15 +26,14 @@ import org.apache.spark.sql.internal.SQLConf
  * If the UDF is deterministic and if the children are all literal, we can replace the udf
  * with the output of the udf serialized
  */
-object DeterministicLiteralUDF extends Rule[LogicalPlan] {
+object DeterministicLiteralUDFFolding extends Rule[LogicalPlan] {
   def apply(plan: LogicalPlan): LogicalPlan =
-    if (!SQLConf.get.deterministicLiteralUdfFoldEnabled) {
+    if (!SQLConf.get.deterministicLiteralUdfFoldingEnabled) {
       plan
     } else plan transformAllExpressions {
-      case udf @ ScalaUDF(_, dataType, children, _, _, _, _, udfDeterministic)
-        if udf.deterministic && children.forall(_.isInstanceOf[Literal]) => {
+      case udf @ ScalaUDF(_, dataType, children, _, _, _, _, _)
+          if udf.deterministic && children.forall(_.isInstanceOf[Literal]) =>
         val res = udf.eval(null)
         Literal(res, dataType)
-      }
     }
 }
